requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
emoji==0.5.0
Flask==0.12.2
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-telegram-bot==10.0.1
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
